Allow build to pass on PHP7 and HHVM.
xDebug is not available in PHP7 and HHVM at the time of this commit.
Since an xDebug function is using as part of the test, this causes the
tests to fail for PHP7 and HHVM.

For now, in the interest of getting tests to pass, let's only run the
xDebug tests when the functions exist. We'll revisit this when xDebug is
a reality for PHP7 and HHVM.