Do not purge cache on "edit_post".
If the cache is purged on "edit_post" and "save_post", the result is
that the purge requests will be doubled. Presumably, the "edit_post"
action was added for invalidating a post after a comment is added. This
is a little heavy handed in that this filter is in multiple places
causing the duplication of requests.