User surrogate key purging for a perf boost.
In testing this in a local dev environment with some dubious network
conditions, purging on save was taking ~30s. This was because 14
requests needed to be made and parallel requests in a WP plugin just
isn't quite possible yet.

Since we have surrogate keys, we should just use those to purge content.
This means that we only issue a single request, and presumably, we do a
much better job purging the right pages.