Default to soft purge
Instead of instant purging by default, the plugin is transitioning to
providing soft purges by default. This decision was made in order to
provide the most reliable experience for various WP installs. Soft purge
is an excellent default experience as it allows for stale content to be
served while a cached is revalidated.

In order to facilitate this change, I reconceptualized the notion of
soft purge and not soft purge into a "purge type" value. You can now
have a "soft" and "instant" purge type. This language is now reflected
in a configuration value, WP CLI, and labelling throughout the code
base. This is a forward thinking refactor in that it allows for support
for future purge types (whether a Fastly implement thing or not).