Update WP CLI command name.
"fp" is a terrible command name. It is not clear at all what it is doing
with that name. In the interest of UX and clarity, the namespace for the
commands is changed to "fastly" and the main command is now "purge".
Having a top level "fastly" command also sets the stage for additional
commands.