Send headers using the `wp` action.
The `wp` action represents the first point in the WordPress execution
in which the context for the request is 100% set. At this point, code
can determine the type of request, access the query object, etc. Is
it sufficient to set headers at this point.

This commit moves the header setting *and* sending to the `wp` action.
The primary issue is that if we wait until `template_redirect` to send
the headers, requests that never get to `template_redirect` will not
have any headers sent. For example, RSS feeds and WP API endpoints
will not have surrogate key headers set. By bumping the action to
`wp`, more WordPress requests will get headers set.

Note that the priority is 101 to ensure that the action that sets
the surrogate keys has run. This will also make it easier for devs
to set their own headers without having to worry too much about
priorities.